requirements.txt
pandas~=2.2.2
opencv-python==4.10.0.84
pytest~=6.2.5
typing~=3.7.4.3
numpy~=1.26.4
pip==24.2
black==23.1.0
isort==5.11.4
pre-commit==2.20.0
torch~=2.6.0
torchvision~=0.21.0
setuptools~=75.1.0
omegaconf~=2.3.0
hydra-core~=1.3.2
google-cloud-aiplatform